chore: rename entry-react-server -> entry-server
